utils/extract-buscos/requirements.txt
biopython==1.83
PyYAML==6.0.1
